+++ src/Router.js
@@ -10,10 +10,13 @@ import MainYuna from "./pages/yuna/Main/Main";
 import LoginTaeyoon from "./pages/taeyoon/Login/Login";
 import MainTaeyoon from "./pages/taeyoon/Main/Main";
 
+import Nav from "./components/Nav/Nav";
+
 // import 한 컴포넌트에 대한 경로를 각각 설정해줍니다.
 const Router = () => {
   return (
     <BrowserRouter>
+      <Nav />
       <Routes>
         <Route path="/login-hyobin" element={<LoginHyobin />} />
         <Route path="/main-hyobin" element={<MainHyobin />} />
